requirements.txt
anaconda==0.0.1.1
asttokens==2.0.5
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
chromedriver-autoinstaller==0.4.0
click==8.1.3
colorama==0.4.4
cryptography==35.0.0
cycler==0.11.0
decorator==5.1.1
executing==0.8.3
Flask==2.2.2
fonttools==4.33.3
gunicorn==20.1.0
h11==0.12.0
heroku==0.1.4
idna==3.3
importlib-metadata==5.0.0
ipython==8.3.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.2
MarkupSafe==2.1.1
mysql-connector==2.2.9
nltk==3.7
outcome==1.1.0
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.1
prompt-toolkit==3.0.29
pure-eval==0.2.2
pycparser==2.21
Pygments==2.12.0
pyOpenSSL==21.0.0
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.1
regex==2022.4.24
requests==2.26.0
schedule==1.1.0
scikit-learn==1.1.1
scipy==1.8.1
selenium==4.0.0
six==1.16.0
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
stack-data==0.2.0
threadpoolctl==3.1.0
tqdm==4.64.0
traitlets==5.2.1.post0
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
wcwidth==0.2.5
webdriver-manager==3.8.3
Werkzeug==2.2.2
wsproto==1.0.0
zipp==3.8.1
